gd/hci: add a msft module to handle Microsoft Extension
This patch adds a msft module to handle Microsoft Extension.

Through the MGMT socket, this module queries the kernel whether
the controller supports the Microsoft extension. If yes, the
Microsoft opcode is saved. With the opcode, the msft module
sends an HCI command to query the Microsoft supported features
and the vendor prefix to handle more Microsoft commands and
events later.

Bug: 246398494
Tag: #floss
Test: ./build.py --target test
Test: Manual test with next patch.
BYPASS_LONG_LINES_REASON: Bluetooth likes 120 lines

Change-Id: Ifb60eeff8c8791f42fcca48bd812886c031cd9ba