Floss: Refactor btif_callbacks_dispatcher
This refactors dispatcher functions like dispatch_base_callbacks to not
assume that the generated dispatcher belongs to a particular struct, but
rather they accept object reference as parameters. This is useful when
we want to apply the dispatcher functions to multiple objects, for
example dispatching LeRand message to Suspend object instead in addition
to Bluetooth object.

Bug: 232547719
Tag: #floss
Test: Build Floss on Linux

Change-Id: Ia68f6be3af33e04201745b4d7601d015dcebe29c