vc: Send control point request only if device is ready
When trying to do control point operations, being connected
is not enough as we also need to know the volume change counter.

If a group API is used to set volume bofore the individual
devices are reported as connected, we should not set it
as we dont know the volume change counter yet. The group volume
which could not be applied because of this, will be applied from
the cache once we report the device as connected.

Test: atest --host bluetooth_vc_test --no-bazel-mode
Bug: 248475583
Tag: #feature
Change-Id: I27b462f7f55122cc7c38fb18908d52e86ea4b5d8
Merged-In: I27b462f7f55122cc7c38fb18908d52e86ea4b5d8
(cherry picked from commit 90b2517650ffd113711a404ed2b4fec694d035fd)