pdl: Generate canonical parsing tests
The new binary in src/bin/generate-canonical-tests.rs will take a JSON
file with canonical test vectors (from tests/canonical/) and produce a
Rust file with unit tests for each test vector.

This allows us to verify the full end-to-end functionality of the
generated Rust code: we check that it compiles and that it gives the
correct values. This change only contains the parsing tests, we will
test serialization next.

Test: atest pdl_rust_generator_tests_{le,be}
Change-Id: I88d7de98e7a11b6c8d2ceffadc4d3282db638a14