vc: Fix doing initial reads and writes twice
We already handle encryption complete event which is more
reliable than passing callback to the enryption request.
We should have a single source of true for the encryption
result. For a reconnecting device having both was the cause
of duplicate GATT operations for the initial reads and writes
(change counter reads etc.)

Bug: 248475583
Tag: #feature
Test: atest --host bluetooth_vc_test --no-bazel-mode, manual testing
Change-Id: Ia4f40a6bead331883a381205cc150e009ec18070