Add Bluetooth API shim compatibility with new SDK versions.
Telecomm is adding a new extra to Call.java in SDK 34, and
to enable Bluetooth to use it and keep compatibility with
older SDK versions we need to implement an API shim that
will select the right method at build time.
This will also be used for each API change in SDK > 33.

The classes in the API shim for the SDK version the
Bluetooth module is built on should have a package name
without a version in it (com.android.bluetooth.apishim)
then each version should check for the platform version
and call the previous shim until it find a suitable one.

Bug: 246953241
Tag: #feature
Test: not yet
Ignore-AOSP-First: New API to shim is only in internal
Change-Id: I508543b893eafed8169b3432cba8ff25364d4621