[avatar_experimental] - Copy avatar from external
Copy avatar from external into pandora and mark it
as experimental and link it with the interfaces in
pandora_experimental.
At some point avatar will be only in external/ but
currently it is easier to iterate on it and more
flexible in avatar_experimental.

Bug: 245578454
Test: atest-dev avatar -v
Ignore-AOSP-First: Cherry-pick from AOSP
Merged-In: I79a2ac247dc37becc61e2fc2158f5836f1894cfe
Change-Id: I79a2ac247dc37becc61e2fc2158f5836f1894cfe