CTKD from LE to Classic stability fix
When bond is established using LE transport with CTKD, we would start a
500ms timer to check if other side reports any errors. If connection or
disconnection happens during that time on classic transport, we would
reset the SMP control block, and cancel this alarm.
Even if it were to fire, the state machine would go into idle state, and
we won't send some callbacks that are crucial to inform upper layers
that the pairing was indeed successfull.

After this patch, classic connection/disconnection doesn't impact the
SMP control block, and the timer always fires.

Bug: 246560805
Test: Bond using LE transport 10 times in a row with device supporting
      CTKD
Change-Id: Id97aa338a20163b1d571c48b4dba224c95ee4880