LE Audio: speed up connection to CSIS devices
When scanning, we can discover multiple devices belonging to same set,
but would trigger pairing to just one of them. We can keep the RSI in
inquiry database, and reuse it after pairing to first device is done, to
speed up pairing with other devices in the set.

It looks like some devices might decide to leave discoverable mode, and
stop advertising RSI when first device in set is bonded, or some short
time after that, while we are still doing service discovery. Without
this patch, we would fail to conenct to other devices in such set.

Test: Pair with LE Audio device, ensure both earbuds are Bonded
Bug: 247504761
Bug: 248340112
Merged-In: I7022bb0dc3422dcdbb141b1c347d43b48d424def
Change-Id: I7022bb0dc3422dcdbb141b1c347d43b48d424def