HFP: No longer crash when service is not available
We can no longer assert the log here.
The service is set as non-available before the native is cleaned up.
Meaning it is not possible to get it, but still the callback can be
called. This only happens during shutdown and adding a synchronized
block between them both will only result in a deadlock.

Fix: 243732562
Change-Id: I3a9c267d2bea083e7529a07f7809e0e444160f80
Test: TH