floss: Migrate device VID:PID and appearance cache
Since aosp/2198218 and aosp/2190680 we have the VID:PID and device
appearance in Floss cache. However, it was not integrated with the
migration script.

This CL migrates those as well.

Note that SdpDi* field might exists in Floss, but it only covers
Classic devices, and therefore are superseded by VendorId* and
Product* fields. It is still kept for compatibility purposes.

Bug: 254227435
Tag: #floss
Test: Manual test and ./build.py --target test
Change-Id: I1e3c77ac882dce6a19d9c7feca2c57860e28a272