floss: Allow HID reconnect during suspend
Raw addresses for paired devices aren't only stored as public addresses.
Since we filter for LE HID attributes, provide both the RawAddress and
LE address type when we provide the list of devices that can wake the
system.

In order for LeAddressManager to re-arm connection, we need to add these
devices via |AcceptLeConnectionFrom| instead of
|AddDeviceToFilterAcceptList|.

Bug: 231345733
Tag: #floss
Test: Manual test on ChromeOS.
Change-Id: I8ae8bc0f71ed07260f5f1880ab6bb8f27ffff972