Add exported flag to exposed test runtime receivers
Android T allows apps to declare a runtime receiver as not exported
by invoking registerReceiver with a new RECEIVER_NOT_EXPORTED flag;
receivers registered with this flag will only receive broadcasts from
the platform and the app itself. However to ensure developers can
properly protect their receivers, all apps targeting a future
platform release and registering a receiver for non-system
broadcasts must specify either the exported or not exported flag
when invoking #registerReceiver; if one of these flags is not
provided, the platform will throw a SecurityException. This commit
updates all the exposed receivers with the RECEIVER_EXPORTED flag
to maintain the existing behavior for these tests.

Bug: 234659204
Tag: #compatibility
Test: atest BassClientServiceTest
Test: atest HapClientTest
Change-Id: Ib832007a0c80755e9a50d029eca45a69b4011ea6