Floss: add more logs to bluetooth_media.rs
There are many steps involved to get a bluetooth audio device stream
A2DP/HFP, and we are missing the logs for some of them to be able to
identify which step went wrong.

In this CL, we add more logs to show the exposed profiles of the
peripheral, and the timings of audio config setup and start/stop
request.

Bug: 254694739
Tag: #floss
Test: Build and verify
Change-Id: Icd375f5a4bfb491c9aad510a4547a56df5775e8d