Floss: add AvrcpController to connect/disconnect labels
We suspect in some extreme cases only AvrcpController is exposed as
an available media profile, but the peripheral still manages to connect
to other profiles.

When the user disconnects the peripheral via UI,
`disconnect_all_enabled_profiles` is invoked. However, since
AvrcpController is not in the media labels, it will not propagate to the
media layer disconnect functions.

In this CL, we add AvrcpController to the labels (also on
`connect_all_enabled_profiles` for symmetry), and add a few logs to help
debugging around SDP.

Bug: 255990901
Tag: #floss
Test: Build and verify
Change-Id: I4f36de2358bf9509110192c26e6bff91371bbd1d