floss: topshim: Deduplicate RawAddress definitions.
RawAddress lived with several names in topshim: RawAddress, FfiAddress,
and RustRawAddress. Sometimes they are properly converted, but sometimes
we just cast. The conversions are actually redundant, and the casting
has no guarantee to be safe.

This patch makes C++ RawAddress the single source of truth, and only the
bindgen-ed RawAddress will be used in the Rust world.

Bug: 249685703
Tag: #floss
Test: ./build.py --target clean && ./build.py --target all && ./build.py --target test
Change-Id: I1f32e61a4c0a019d03a25e5ad63c3dd0e038f460