Move all profile services bind to BluetoothManagerService
We are currently binding the profile services in framework/
BluetoothProfileConnector, except for BluetoothHeadset
and BluetoothLeCallControl that are bound in BMS.

The issue with binding the profiles in framework/ is that
calling apps don't always have the required permission
(BLUETOOTH_PRIVILEGED) to check if the service is enabled
or not (BluetoothAdapter#getSupportedProfiles), so we are
trying to bind the service even if the profile is disabled.

This change will unify the profiles service binding by
moving the bind process in BluetoothProfileConnector in
BluetoothManagerService and making BluetoothHeadset and
BluetoothLeCallControl use BluetoothProfileConnector.

Bug: 241827236
Bug: 236399693
Test: atest CtsBluetoothTestCases
Test: atest BluetoothInstrumentationTests
Tag: #feature
Ignore-AOSP-First: Cherry-pick
Change-Id: I925b4415c023ff9e43a8ff947ae7ca5456babe64
Merged-In: I925b4415c023ff9e43a8ff947ae7ca5456babe64