LeAudio: Adjust audio set configurations
This changes the "Recording" scnenario to "Live" for the
consistency with the context type mapping in the configuration
provider implementation. This also better represents the actual
target use case which will be bidirectional and not just for the
microphone.

Bug: 233976574
Test: manual testing
Change-Id: I84391473d6220261eb2ecf50f10fe60866c8fb52
Merged-In: I84391473d6220261eb2ecf50f10fe60866c8fb52
(cherry picked from commit 6208edb684de67eb9bfd8ff154f25614913cac01)