LeAudio: Remove inconsistency in BluetoothAdatper.getActiveDevices
UI expects to get all the devices which are in the active group.
On the other side, LeAudioService return OUT/IN active device which was
not used by the application.
With this patch, BluetoothLeAudio.getActiveDevices() will return devices
which belongs to the Active group, and first element is the Lead device.

Bug: 248188703
Test: atest BluetoothInstrumentationTests
Tag: #feature
Change-Id: I84a5d5871f20eed5ef0a07291c8c28787043d095