Register BT receiver in ActiveDeviceManager as exported
The receiver in ActiveDeviceManager was recently updated to also
receive HAP_CONNECTION_STATE_CHANGED broadcasts; however, since this
action is not currently tagged as a protected-broadcast, this
receiver must be flagged with the exported state. Since the
bluetooth actions cannot be delivered to an unexported receiver,
this commit flags this receiver as exported.

Bug: 234659204
Tag: #compatibility
Test: atest ActiveDeviceManagerTest
Change-Id: Ie8e81ae905dbba45d8b2dc2449aeecd0128be815