[Pandora] - Simulate CSR command timeout
Bug: 259102046
Test: atests pts-bot:AVRCP -v

Introducing a 2s sleep at the beggining of the avrcp tests to Simulate
the previous non implemented rootcanal command vendor.

The CSR vendor command was not implemented and timed out at the
beggining of each HFP tests. Since they are implemented in rootcanal,
the timeout doesn't exists and the tests are failing.

Change-Id: I6838149dd1f34477f48df3ae767656a29a4aa678