Fix BluetoothOppManagerTest
The URI that is used for test is only valid in the local test device.
Therefore the test passed locally even without mocking contentResolverInsert(). However, it should fail since there is no content for the URI in other devices.

This CL fixes the bug by making the contentResolverInsert() do nothing.

Bug: 260183711
Tag: #refactor
Test: atest BluetoothOppManagerTest
Change-Id: I0ba90c1236cd846db4fbf2abb233a9b679274657