LeAudio: Add timeout to VBC close
Once the Voiceback channel is closed (recording is suspended)
we usually reconfigure to a different configuration on the next
playback track metadata update. This change adds an additional
trigger if no playback track metadata is updated after the
recording has finished. This is just to prepare for a more common
MEDIA playback scenario before any playback event happens, which
should reduce the initial configuration time in most common
use cases.

Bug: 256967923
Test: manual testing
Change-Id: I1b09f805c06eb918d03b4bf67a3d26fcb0513569