Disable OPUS codec when using A2DP HIDL HAL.
On bramble and likely other phones that use A2DP HIDL HAL, when
connected to Pixel Buds Pro, we try to select the OPUS codec as it is
the preferred codec for Pixel Buds Pro. However, since OPUS support was
added to Android after A2DP HIDL HAL freeze, we get an error from the
HIDL HAL.

While initializing codecs, check if we are using the A2DP HIDL HAL and
disable OPUS codec if that is true.

Bug: 254231581
Tag: #refactor
Test: Manually tested using barbet which also has using the A2DP HIDL
HAL
Ignore-AOSP-First: OPUS codec support does not exist on AOSP yet

Change-Id: I1bf9d7a6121d6d6d13098ee38bebb363e444ebaf