leaudio: Fix improper cis calculations for TWS style earbuds
With this patch code takes number of devices in the group in order to
choose best configuration when calculating required CISes.
Also it fixes how strategy is taken into account - as for now we want to
check only strategy for sink direction as this is defines group
strategy.

Also fixes unit tests and improves some important logs

Bug: 261624834
Test: atest BluetoothInstrumentationTests
Test: atest bluetooth_le_audio_test bluetooth_le_audio_client_test
Tag: #feature

Change-Id: I00a0cc426d193a78938609fce0e620a55d9102e1