Call setActiveDevice(null) when disconnected
When A2DP is disconnected in call mode, and when HFP is disconnected in
media mode, the disconnection of the audio device is not propagated to
the A2dpService or HeadsetService via setActiveDevice(null).
Thus, when the same audio device reconnected after, the audio is not came
from the audio device.

Bug: 261366232
Tag: #refactor
Test: atest BluetoothInstrumentationTests
Change-Id: Ic80b4ac436d61f0c8129e9bd61e086fe4c5d4644