Floss: Implement commands for socket test schedule
In qualification, tests in L2CAP/* require DUT to send data or
verify the received data.
Since the L2CAP connection can be disconnected due to a short time out,
this CL implements some help schedules to automatically send and dump
the data for the users to run qualification tests.

Bug: 262813906
Test: run L2CAP qualification tests locally

Change-Id: Ia00953460959eb60ac86b81149235f87ea6bedcc