[RESTRICT AUTOMERGE] Reduce behavior change with LE ACL timeout on T
See previous CL for details. Here we constrain the change to when an
ATT control block is already allocated, so the behavior with is
*exactly* the same except for with LE CoC (where the bugfix is added).

But in U, for consistency, we might as well have the same behavior in
classic / in general, to make it more predictable. So this patch will
only go in for T-QPR2.

Ignore-AOSP-First: Only needed in QPR
Bug: 251954527
Test: atest pts-bot
Change-Id: I5f065d064fb8a9a67861811d0559f46115571580