Revert "Set connection state to not connected while ACL is lost"
This reverts commit b9dd9b84e37bdaa788393619786a92ebe89ed1ab.

Reason for revert: This is not valid fix.
device->conn_state is written when device might be not valid anymore, as
elements in the bta_dm_cb.device_list.peer_device[] are moved before device->conn_state is written.
Also, code already clears the  disconnected peer_device, so the patch is invalid.

It also might lead to situation, that conn_state for the connected device is incorrectly overriten.

Bug: 263698493
Change-Id: Id58a371a447c6f2ac4ad35dabdda2d73ca8fe603
Test: manual connected disconnect set members and verify device.isConnected()
Tag: #feature