leaudio: Update confiuration, when device connected
With this patch, we make sure, device is taken to the audio
configuration, when it is connected on the service level.
Othwerwise, we could end up with not being able to create stream,
because configuration already includes the device which is not ready.
Usually can happen with late device connect.

Bug: 262309077
Test: atest BluetoothInstrumentationTest
Test: manual streaming with late connect
Tag: #feature
Change-Id: I0698a429b7ca6060e6fba57dbe59207ca658abde