floss: Expose debug logging to runtime
In order to improve debugability, we should be able to control the log
level without having to restart the daemon. Add a d-bus method to get
and set debug logging. Persistence will be handled in a subsequent CL
with a similar api in the manager daemon.

NO_IFTTT=no change required on flag dump

Bug: 241963222
Tag: #floss
Test: Manual test on ChromeOS
Change-Id: Ib95c6dd989c20bd5a1f7eba5dab1887865d1f5d5